change how markov state is stored on disk
now we only store a std::map keyed with the hash of the prefix,
instead of a hashmap of the entire hashmap. while this means that
there might be collisions, markov isn't critical anyway.

this also means that we lose the ability to "explore" the markov
table, but given the sheer number of entries currently (800k),
it's not very feasible in any case, so it doesn't matter that much